disable auto rehash for mysql client
MySQL scans database, tables and columns name for auto completion. If you have many db, tables it could take a while. To avoid that, launch your client with the -A option (or --no-auto-rehash). You could also add the disable_auto_rehash variable in your my.cnf (in the [mysql] section) if you want to disable it completely.